requirements.txt
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
httplib2==0.22.0
idna==3.10
numpy==2.2.6
opencv-python==4.12.0.88
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
requests==2.32.4
RPi.GPIO==0.7.1
rsa==4.9.1
screeninfo==0.8.1
uritemplate==4.2.0
urllib3==2.5.0
